feat: Move builder url func
